Add opts.classMap to customize class names generated by uPlot
classMap is a function taking a class name and returning a class name. Apply
this function to every class name generated by uPlot.

This fixes one issue when using uPlot with Bulma.css. Bulma has a global .select
class that interferes with uPlot's .select class. The effect is an annoying blue
"drop-down" symbol drawn on the plot area.

The default for opts.classMap is the identity map.