Update README.md exponential -> linear
let's say you have 10 datasets with 100k entries each. 

this is 1.1M input data points spread out across 11 arrays holding 100k entries each (1 for x, 10 for y's)

if the situation changes from sharing x values to all disjoint x-values, there will be 1M unique x values, and y arrays will be padded with 9x as many nulls as there are actual data. At all times x and y arrays must have the same length.

This results in 11M array values in the data in order to represent this. This is just a linear proportion to the number of independent x-values here not a quadratic or exponential increase. 

If we cared about making the representation more efficient here for memory usage, it should not be hard to add a separate convention to put multiple datasets into one graph, that way if you could group the x and y values that are shared together.

In this example, then there would be 10 datasets each with x and y values, making for 20 arrays of 100k elements = 2M scalars.