got rid of "wide" links in index.html
The anchor tags on index.html page were styled to be block elements (display: block), that caused the entire width of the page around a given anchor to be clickable... which has repeatedly annoyed me when trying to look at the demos / learn uplot.

Example: when focusing a browser window displaying the index.html by clicking in whitespace on the page, you accidentally trigger the link to whatever demo your pointer happens to be horizontally aligned with.

I removed the display:block from the anchor tags, wrapped said anchors in <div>, and gave each div the class 'demo'.  The page looks the same, but the links are now confined to the immediate space around the anchors.